Session 4 - IoT Digital Twin/Functions/iothub_import/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-digitaltwins-core 
azure-identity

